pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "freeact-skills"
version = "0.0.1"
description = "Library of predefined freeact agent skills"
homepage = "https://github.com/gradion-ai/freeact-skills"
readme = "README.md"
license = "Apache-2.0"
authors = [
    "Martin Krasser <martin@gradion.ai>",
    "Christoph Stumpf <chris@gradion.ai>",
]
packages = [
    { include = "freeact_skills" },
]

[tool.poetry.dependencies]
beautifulsoup4 = { version = "^4.12", optional = true }
google-genai = { version = "^0.2.2", optional = true }
httpx = { version = "^0.27.2", optional = true }
openai = { version = "^1.56.0", optional = true }
pydantic = { version = "^2.9", optional = true }
python = "^3.10,<3.13"
python-dateutil = { version = "^2.9", optional = true }
pyyaml = { version = "^6.0", optional = true }
pyzotero = { version = "^1.5", optional = true }
requests = { version = "^2.32", optional = true }

[tool.poetry.extras]
reader = ["python-dateutil", "requests"]
resume = ["beautifulsoup4", "pydantic", "pyyaml", "requests"]
search-google = ["google-genai"]
search-perplexity = ["openai", "httpx"]
zotero = ["python-dateutil", "pyzotero"]
all = [
    "beautifulsoup4",
    "google-genai",
    "httpx",
    "openai",
    "pydantic",
    "python-dateutil",
    "pyyaml",
    "pyzotero",
    "requests"
]

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0"
invoke = "^2.2"
pytest = "^8.3"
pytest-asyncio = "^0.24.0"
python-dotenv = "^1.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
line-length = 120

[tool.ruff.lint.per-file-ignores]
"**/__init__.py" = ["F401"]

[tool.mypy]
python_version = 3.11
check_untyped_defs = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "dateutil.relativedelta",
    "dateutil.parser",
    "requests",
    "yaml"
]
ignore_missing_imports = true
